fix(agent): Rework code to reduce private includes required
